+++ graph-parser_c/src/graph/graph.h
@@ -21,11 +21,11 @@ extern "C" {
     struct node_graph {
         struct list neighbours;
         uint8_t name;
-        int node_graph_id;
+        uint8_t node_graph_id;
     }; 
     struct edge_graph {
         struct node_graph * to;
-        double value;
+        float value;
         
     }; 
     struct graph {
@@ -33,8 +33,8 @@ extern "C" {
     };
     
     void init_graph(struct graph * g);
-    void add_edge_graph(struct graph * g,  uint8_t name_from,  uint8_t name_to, double value, bool directed);
-    void add_edge_graph_return_node_indexes(struct graph * g,  uint8_t name_from,  uint8_t name_to, double value, bool directed,int  * nodefrom, int * nodeto);
+    void add_edge_graph(struct graph * g,  uint8_t name_from,  uint8_t name_to, float value, bool directed);
+    void add_edge_graph_return_node_indexes(struct graph * g,  uint8_t name_from,  uint8_t name_to, float value, bool directed,uint8_t  * nodefrom, uint8_t * nodeto);
     void free_graph(struct graph * g);
